backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
PyMySQL==1.1.1
cryptography==42.0.8
opencv-python==4.10.0.84
pytesseract==0.3.13
openpyxl==3.1.5
